# Search input performance improvement