--- a/src/vuejs/app/src/components/InputForm.vue
+++ b/src/vuejs/app/src/components/InputForm.vue
@@ -1,0 +1,35 @@
+<template>
+  <b-form @submit="submit">
+    <b-input-group>
+      <b-form-input v-model="searchQuery"
+                    :disabled="disabled"
+                    :placeholder="placeholder">
+      </b-form-input>
+      <b-input-group-append>
+        <b-btn type="submit"
+               variant="primary">Search</b-btn>
+      </b-input-group-append>
+    </b-input-group>
+  </b-form>
+</template>
+
+<script>
+export default {
+  name: 'InputForm',
+  props: {
+    disabled: Boolean,
+    placeholder: String
+  },
+  data() {
+    return {
+      searchQuery: ''
+    }
+  },
+  methods: {
+    submit(e) {
+      e.preventDefault()
+      this.$emit('submit', this.searchQuery)
+    }
+  }
+}
+</script>