--- a/src/vuejs/app/src/components/SearchForm.vue
+++ b/src/vuejs/app/src/components/SearchForm.vue
@@ -15,7 +15,7 @@
 
 <script>
 export default {
-  name: 'InputForm',
+  name: 'SearchForm',
   props: {
     disabled: Boolean,
     placeholder: String