This PR is search input performance improvement more faster.
After improvement, rendered at 60 fps while typing.

|Before|After|
|-|-|
|<img width="558" src="https://user-images.githubusercontent.com/5723074/59889390-1c2e7780-9407-11e9-8e68-68f865a50da0.png">|<img width="561" alt="image" src="https://user-images.githubusercontent.com/5723074/59889451-6f082f00-9407-11e9-9e81-831b06e7283a.png">|