@tekezo 2019-06-21
Your awesome improvements has been deployed!
https://pqrs.org/osx/karabiner/complex_modifications/